+++ src/GameData/UserData.cs
@@ -35,6 +35,16 @@ namespace PrepareLanding.GameData
         public UsableMinMaxNumericItem<float> AverageTemperature { get; } = new UsableMinMaxNumericItem<float>();
 
         /// <summary>
+        ///     Current user choices for the minimum temperature.
+        /// </summary>
+        public UsableMinMaxNumericItem<float> MinTemperature { get; } = new UsableMinMaxNumericItem<float>();
+
+        /// <summary>
+        ///     Current user choices for the maximum temperature.
+        /// </summary>
+        public UsableMinMaxNumericItem<float> MaxTemperature { get; } = new UsableMinMaxNumericItem<float>();
+
+        /// <summary>
         ///     Current user choice for the "Animal Can Graze Now" state.
         /// </summary>
         public MultiCheckboxState ChosenAnimalsCanGrazeNowState
@@ -130,10 +140,29 @@ namespace PrepareLanding.GameData
             }
         }
 
+        public ThingDef ForagedFood
+        {
+            get => _foragedFood;
+
+            set
+            {
+                if (value == _foragedFood)
+                    return;
+
+                _foragedFood = value;
+                OnPropertyChanged(nameof(ForagedFood));
+            }
+        }
+
         /// <summary>
-        ///     Current user choices for the current movement time.
+        ///     Current user choices for forageability.
         /// </summary>
-        public UsableMinMaxNumericItem<float> CurrentMovementTime { get; } = new UsableMinMaxNumericItem<float>();
+        public UsableMinMaxNumericItem<float> Forageability { get; } = new UsableMinMaxNumericItem<float>();
+
+        /// <summary>
+        ///     Current user choices for movement difficulty.
+        /// </summary>
+        public UsableMinMaxNumericItem<float> MovementDifficulty { get; } = new UsableMinMaxNumericItem<float>();
 
         /// <summary>
         ///     Current user choices for the elevation.
@@ -222,29 +251,10 @@ namespace PrepareLanding.GameData
             TileFilterCoastRotation.PossibleRotationsInt, TileFilterCoastRotation.PossibleRotationsInt[0]);
 
         /// <summary>
-        ///     Current user choices for the summer movement time.
-        /// </summary>
-        public UsableMinMaxNumericItem<float> SummerMovementTime { get; } = new UsableMinMaxNumericItem<float>();
-
-        /// <summary>
-        ///     Current user choices for the summer temperature.
-        /// </summary>
-        public UsableMinMaxNumericItem<float> SummerTemperature { get; } = new UsableMinMaxNumericItem<float>();
-
-        /// <summary>
         ///     Current user choices for the time zone.
         /// </summary>
         public UsableMinMaxNumericItem<int> TimeZone { get; } = new UsableMinMaxNumericItem<int>();
 
-        /// <summary>
-        ///     Current user choices for the winter movement time.
-        /// </summary>
-        public UsableMinMaxNumericItem<float> WinterMovementTime { get; } = new UsableMinMaxNumericItem<float>();
-
-        /// <summary>
-        ///     Current user choices for the winter temperature.
-        /// </summary>
-        public UsableMinMaxNumericItem<float> WinterTemperature { get; } = new UsableMinMaxNumericItem<float>();
 
         /// <summary>
         ///     Current user choice for Most / Least item
@@ -295,13 +305,13 @@ namespace PrepareLanding.GameData
             if (_stoneTypesNumberOnly)
                 return false;
 
-            if (CurrentMovementTime.Use)
+            if (MovementDifficulty.Use)
                 return false;
 
-            if (SummerMovementTime.Use)
+            if (Forageability.Use)
                 return false;
 
-            if (WinterMovementTime.Use)
+            if (_foragedFood != null)
                 return false;
 
             if (Elevation.Use)
@@ -313,10 +323,10 @@ namespace PrepareLanding.GameData
             if (AverageTemperature.Use)
                 return false;
 
-            if (WinterTemperature.Use)
+            if (MinTemperature.Use)
                 return false;
 
-            if (SummerTemperature.Use)
+            if (MaxTemperature.Use)
                 return false;
 
             if (GrowingPeriod.Use)
@@ -352,6 +362,7 @@ namespace PrepareLanding.GameData
             _coastalLakeTileState = MultiCheckboxState.Partial;
             CoastalRotation.Reset(false);
             _chosenAnimalsCanGrazeNowState = MultiCheckboxState.Partial;
+            _foragedFood = null;
 
             var defProps = PrepareLanding.Instance.GameData.DefData;
             SelectedRoadDefs.SetContainer(defProps.RoadDefs, nameof(SelectedRoadDefs));
@@ -363,9 +374,8 @@ namespace PrepareLanding.GameData
             StoneTypesNumber = 2;
 
             // min / max numeric fields containers
-            InitUsableMinMaxNumericItem(CurrentMovementTime, nameof(CurrentMovementTime));
-            InitUsableMinMaxNumericItem(SummerMovementTime, nameof(SummerMovementTime));
-            InitUsableMinMaxNumericItem(WinterMovementTime, nameof(WinterMovementTime));
+            InitUsableMinMaxNumericItem(MovementDifficulty, nameof(MovementDifficulty));
+            InitUsableMinMaxNumericItem(Forageability, nameof(Forageability));
             InitUsableMinMaxNumericItem(Elevation, nameof(Elevation));
             InitUsableMinMaxNumericItem(TimeZone, nameof(TimeZone));
 
@@ -377,8 +387,8 @@ namespace PrepareLanding.GameData
              */
 
             InitUsableMinMaxNumericItem(AverageTemperature, nameof(AverageTemperature));
-            InitUsableMinMaxNumericItem(WinterTemperature, nameof(WinterTemperature));
-            InitUsableMinMaxNumericItem(SummerTemperature, nameof(SummerTemperature));
+            InitUsableMinMaxNumericItem(MinTemperature, nameof(MinTemperature));
+            InitUsableMinMaxNumericItem(MaxTemperature, nameof(MaxTemperature));
 
             var twelfthList = Enum.GetValues(typeof(Twelfth)).Cast<Twelfth>().ToList();
             GrowingPeriod =
@@ -469,6 +479,9 @@ namespace PrepareLanding.GameData
         ///     The currently selected hilliness state.
         /// </summary>
         private Hilliness _chosenHilliness;
+
+
+        private ThingDef _foragedFood;
 
         /// <summary>
         ///     If true, filter only tiles with only a given number of stone types.