+++ src/Presets/Preset.cs
@@ -49,9 +49,8 @@ namespace PrepareLanding.Presets
             _gameData.UserData.ChosenHilliness = LoadEnum<Hilliness>(xTerrain, "Hilliness");
             LoadThreeStateItemContainer(xTerrain, "Roads", "Road", _gameData.UserData.SelectedRoadDefs);
             LoadThreeStateItemContainer(xTerrain, "Rivers", "River", _gameData.UserData.SelectedRiverDefs);
-            LoadUsableMinMax(xTerrain, "CurrentMovementTime", _gameData.UserData.CurrentMovementTime);
-            LoadUsableMinMax(xTerrain, "SummerMovementTime", _gameData.UserData.SummerMovementTime);
-            LoadUsableMinMax(xTerrain, "WinterMovementTime", _gameData.UserData.WinterMovementTime);
+            LoadUsableMinMax(xTerrain, "MovementDifficulty", _gameData.UserData.MovementDifficulty);
+            LoadUsableMinMax(xTerrain, "Forageability", _gameData.UserData.Forageability);
             if (xTerrain.Element("StoneTypesNumberOnly") == null)
             {
                 LoadThreeStateItemContainerOrdered(xTerrain, "Stones", "Stone", _gameData.UserData.SelectedStoneDefs);
@@ -80,8 +79,8 @@ namespace PrepareLanding.Presets
                 return;
 
             LoadUsableMinMax(xTemperature, "AverageTemperature", _gameData.UserData.AverageTemperature);
-            LoadUsableMinMax(xTemperature, "SummerTemperature", _gameData.UserData.SummerTemperature);
-            LoadUsableMinMax(xTemperature, "WinterTemperature", _gameData.UserData.WinterTemperature);
+            LoadUsableMinMax(xTemperature, "MinTemperature", _gameData.UserData.MinTemperature);
+            LoadUsableMinMax(xTemperature, "MaxTemperature", _gameData.UserData.MaxTemperature);
             LoadMinMaxFromRestrictedList(xTemperature, "GrowingPeriod", _gameData.UserData.GrowingPeriod);
             LoadUsableMinMax(xTemperature, "RainFall", _gameData.UserData.RainFall);
             _gameData.UserData.ChosenAnimalsCanGrazeNowState = LoadThreeState(xTemperature, "AnimalsCanGrazeNow");
@@ -147,9 +146,8 @@ namespace PrepareLanding.Presets
                 SaveHilliness(xTerrainFilters, "Hilliness", _gameData.UserData.ChosenHilliness);
                 SaveThreeStateItemContainer(xTerrainFilters, "Roads", "Road", _gameData.UserData.SelectedRoadDefs);
                 SaveThreeStateItemContainer(xTerrainFilters, "Rivers", "River", _gameData.UserData.SelectedRiverDefs);
-                SaveUsableMinMax(xTerrainFilters, "CurrentMovementTime", _gameData.UserData.CurrentMovementTime);
-                SaveUsableMinMax(xTerrainFilters, "SummerMovementTime", _gameData.UserData.SummerMovementTime);
-                SaveUsableMinMax(xTerrainFilters, "WinterMovementTime", _gameData.UserData.WinterMovementTime);
+                SaveUsableMinMax(xTerrainFilters, "MovementDifficulty", _gameData.UserData.MovementDifficulty);
+                SaveUsableMinMax(xTerrainFilters, "Forageability", _gameData.UserData.Forageability);
                 if (_gameData.UserData.StoneTypesNumberOnly)
                 {
                     SaveBoolean(xTerrainFilters, "StoneTypesNumberOnly", _gameData.UserData.StoneTypesNumberOnly);
@@ -173,8 +171,8 @@ namespace PrepareLanding.Presets
                 xFilter.Add(xTemperatureFilters);
 
                 SaveUsableMinMax(xTemperatureFilters, "AverageTemperature", _gameData.UserData.AverageTemperature);
-                SaveUsableMinMax(xTemperatureFilters, "SummerTemperature", _gameData.UserData.SummerTemperature);
-                SaveUsableMinMax(xTemperatureFilters, "WinterTemperature", _gameData.UserData.WinterTemperature);
+                SaveUsableMinMax(xTemperatureFilters, "MinTemperature", _gameData.UserData.MinTemperature);
+                SaveUsableMinMax(xTemperatureFilters, "MaxTemperature", _gameData.UserData.MaxTemperature);
                 SaveMinMaxFromRestrictedList(xTemperatureFilters, "GrowingPeriod", _gameData.UserData.GrowingPeriod);
                 SaveUsableMinMax(xTemperatureFilters, "RainFall", _gameData.UserData.RainFall);
                 SaveThreeState(xTemperatureFilters, "AnimalsCanGrazeNow", _gameData.UserData.ChosenAnimalsCanGrazeNowState);